Refactor: Use shared parents to reduce duplication
