+++ lib/frontier/controller_spec/destroy_action.rb
@@ -1,12 +1,9 @@
-class Frontier::ControllerSpec::DestroyAction
-
-  include Frontier::ModelConfigurationProperty
+class Frontier::ControllerSpec::DestroyAction < Frontier::ControllerSpec::MemberAction
 
   def to_s
     raw = <<STRING
 describe 'DELETE destroy' do
-#{render_with_indent(1, subject_block)}
-#{render_with_indent(1, Frontier::FeatureSpec::TargetObjectLetStatement.new(model_configuration).to_s)}
+#{render_with_indent(1, render_setup)}
 
   authenticated_as(:admin) do
     it "deletes the #{model_configuration.as_title}" do
